Properly handle error channel if rotation is used.
When configuring Log::Agent::Driver::File with only:
    -channels => {
	    error => ['server.log', $policy],
    },

the defaults for the other channels were not properly set up since
we used the 'error' channel definition value but forgot to treat that
value as a rotation specification, when given as an ARRAY reference.

This resulted in a filename being the string representation of the
reference, not something pretty :-), and of course no rotation being
used at all for other channels.